Fix validation methods
