+++ sdmetrics/single_table/data_augmentation/utils.py
@@ -1,6 +1,6 @@
 """Utils method for data augmentation metrics."""
 
-from sdmetrics._utils_metadata import _process_data_with_metadata, _validate_single_table_metadata
+from sdmetrics._utils_metadata import _validate_single_table_metadata
 from sdmetrics.single_table.utils import (
     _validate_classifier,
     _validate_data_and_metadata,
@@ -70,14 +70,3 @@ def _validate_inputs(
             'and synthetic data. The following values are present in the synthetic data and'
             f" not the real data: '{to_print}'"
         )
-
-
-def _process_data_with_metadata_ml_efficacy_metrics(
-    real_training_data, synthetic_data, real_validation_data, metadata
-):
-    """Process the data for ML efficacy metrics according to the metadata."""
-    real_training_data = _process_data_with_metadata(real_training_data, metadata, True)
-    synthetic_data = _process_data_with_metadata(synthetic_data, metadata, True)
-    real_validation_data = _process_data_with_metadata(real_validation_data, metadata, True)
-
-    return real_training_data, synthetic_data, real_validation_data
